refactor: 이메일 인증 로직 리팩토링
- 클래스를 생성하고 나눠 역할 분담(인증코드 저장, 이메일 인증 보내기)
