+++ src/main/java/com/dangdang/check/core/employee/EmployeeJpaRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 public interface EmployeeJpaRepository extends JpaRepository<EmployeeEntity, Long> {
 
     Optional<EmployeeEntity> findByLoginIdAndIsDeletedFalse(String loginId);
-    Optional<EmployeeEntity> findByLoginIdAndPasswordAndIsDeletedFalse(String loginId, String password);
+    boolean existsByEmail(String email);
 
 }
