+++ src/main/java/com/dangdang/check/domain/verification/AuthenticationApiService.java
@@ -1,0 +1,6 @@
+package com.dangdang.check.domain.verification;
+
+public interface AuthenticationApiService {
+
+    boolean sendVerificationCode(String email);
+}
