+++ src/main/java/com/dangdang/check/domain/verification/EmailVerificationService.java
@@ -2,5 +2,5 @@ package com.dangdang.check.domain.verification;
 
 public interface EmailVerificationService {
 
-    boolean sendVerificationCode(String email);
+    String createAndSaveCode(String email);
 }
