+++ src/main/java/com/dangdang/check/domain/verification/EmailVerificationServiceImpl.java
@@ -1,6 +1,5 @@
 package com.dangdang.check.domain.verification;
 
-import com.dangdang.check.common.constant.VerificationMessageConstants;
 import com.dangdang.check.common.util.CodeGenerator;
 import com.dangdang.check.core.authentication.EmailVerificationCommandService;
 import lombok.RequiredArgsConstructor;
@@ -14,18 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class EmailVerificationServiceImpl implements EmailVerificationService {
 
     private final EmailVerificationCommandService emailVerificationCommandService;
-    private final EmailSenderService emailSenderService;
 
     @Override
     @Transactional
-    public boolean sendVerificationCode(String email) {
+    public String createAndSaveCode(String email) {
         String code = CodeGenerator.generateAlphaNumericCode();
         emailVerificationCommandService.save(EmailVerificationFactory.from(email, code));
-        emailSenderService.send(email, VerificationMessageConstants.SUBJECT, String.format(VerificationMessageConstants.BODY_TEMPLATE, code))
-                .exceptionally(ex -> {
-                    log.error("Failed to send email", ex);
-                    return false;
-                });
-        return true;
+        return code;
     }
 }
